Add state and move html to appropriate components. Pass all cypress tests.
